Backout conversion of public app to swagger because of increased error rate.
