+++ docs/source/autodocs/auslib.web.views.rst
@@ -1,10 +1,10 @@
-auslib.web.public.views
+auslib.web.views
 ================
 
-auslib.web.public.views.client
+auslib.web.views.client
 -----------------------
 
-.. automodule:: auslib.web.public.views.client
+.. automodule:: auslib.web.views.client
     :members:
     :undoc-members:
     :show-inheritance:
